+++ src/k8s/pkg/k8sd/api/worker.go
@@ -15,6 +15,7 @@ import (
 	"github.com/canonical/k8s/pkg/utils"
 	"github.com/canonical/lxd/lxd/response"
 	"github.com/canonical/microcluster/v2/state"
+	apierrors "k8s.io/apimachinery/pkg/api/errors"
 )
 
 func (e *Endpoints) postWorkerInfo(s state.State, r *http.Request) response.Response {
@@ -59,8 +60,13 @@ func (e *Endpoints) postWorkerInfo(s state.State, r *http.Request) response.Resp
 		return response.InternalError(fmt.Errorf("kubernetes endpoints not ready yet: %w", err))
 	}
 
-	if err := client.CheckNodeNameAvailable(r.Context(), workerName); err != nil {
-		return response.BadRequest(fmt.Errorf("worker node name %s is invalid: %w", workerName, err))
+	// Check if the node name already exists in the cluster.
+	_, err = client.GetNode(r.Context(), workerName)
+	if err == nil {
+		return response.BadRequest(fmt.Errorf("node name already exists: %s", workerName))
+	} else if !apierrors.IsNotFound(err) {
+		// Request to fetch node failed for some other reason
+		return response.InternalError(fmt.Errorf("failed to check whether worker node name is available %s: %w", workerName, err))
 	}
 
 	servers, err := client.GetKubeAPIServerEndpoints(r.Context())
